+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 5.9
 - Do not update Moodle user password in General_Info.inc.php, Student.php, User.php & Moodle/functions.php
 - Hide Last Login on Create Account and Add screens in General_Info.inc.php
 - Add CREATE_STUDENT_ACCOUNT_AUTOMATIC_ACTIVATION to CONFIG table in Update.fnc.php & rosariosis.sql
+- Automatic Student Account Activation in Configuration.php, Student.php & General_Info.inc.php
 
 Changes in 5.8
 --------------